Made some UI changes
Made a couple of UI changes and and made state management more efficient

+++ lib/page_specific_widgets/completed_tasks_page_appbar.dart
@@ -1,5 +1,9 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:simple_todo_app/core/common/show_dialog.dart';
+import 'package:simple_todo_app/core/common/show_error_popUps.dart';
+import 'package:simple_todo_app/core/common/show_snackbar.dart';
 import 'package:simple_todo_app/hive_database/completed_tasks_db_services.dart';
 
 class CompletedTasksPageAppbar extends StatelessWidget
@@ -18,8 +22,33 @@ class CompletedTasksPageAppbar extends StatelessWidget
       actions: [
         IconButton(
           onPressed: () {
-            Provider.of<CompletedTasksDbServices>(context, listen: false)
-                .deleteAllCompletedTasks();
+            if (Provider.of<CompletedTasksDbServices>(context, listen: false)
+                .allCompletedTasks
+                .isNotEmpty) {
+              showConfirmationDialog(
+                context: context,
+                icon: Icons.question_mark,
+                titleText: "Are you sure?",
+                contentText:
+                    "You are about to delete all your completed tasks. Are you sure you want to continue?",
+                onConfirm: () {
+                  Provider.of<CompletedTasksDbServices>(context, listen: false)
+                      .deleteAllCompletedTasks();
+                  showSnackBar(
+                    context,
+                    "Deleted all completed tasks",
+                    Colors.red.shade400,
+                  );
+                  Navigator.of(context).pop();
+                },
+              );
+            } else {
+              showErrorDialog(context,
+                  titleText: "No Completed Tasks",
+                  contentText: "You do not have any completed tasks yet.",
+                  icon: CupertinoIcons.exclamationmark,
+                  actionButtonText: "Complete Tasks");
+            }
           },
           icon: Icon(
             Icons.delete_sweep_rounded,
